// src/types/Membro.ts
export type Membro = {
  name: string
  nacionalidade: string
  dataDeNascimento: string
  naturalidade: string
  fatherName: string
  motherName: string
  estadoCivil: string
  spouseName: string
  cpf: string
  cargo: string
  cep: string
  endereco: string
  bairro: string
  municipio: string
  unidadeFederal: string
  igrejaDeBatismo: string
  dataDeBatismo: string
}